allow functions as content in instructions subsections block.  soundcheck block init method returns container div.
this allows a soundcheck block to be embedded within a subsection of the instructions.  Pass an anonymous function as the subsection content (or as one in a list of contents) which calls the .init() method on the soundcheck block, and the soundcheck div will be embedded in the specified part of the instructions.  Need to debug validation and dynamics behavior though.